+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## v4.0.0-beta.13...v4.0.0-beta.14
+
+[compare changes](https://github.com/LouisMazel/maz-ui/compare/v4.0.0-beta.13...v4.0.0-beta.14)
+
+**Note:** No relevant commits found
+
+
+## v4.0.0-beta.12...v4.0.0-beta.13
+
+[compare changes](https://github.com/LouisMazel/maz-ui/compare/v4.0.0-beta.12...v4.0.0-beta.13)
+
+**Note:** No relevant commits found
+
+
+## v4.0.0-beta.11...v4.0.0-beta.12
+
+[compare changes](https://github.com/LouisMazel/maz-ui/compare/v4.0.0-beta.11...v4.0.0-beta.12)
+
+**Note:** No relevant commits found
+
+
 ## v4.0.0-beta.10...v4.0.0-beta.11
 
 [compare changes](https://github.com/LouisMazel/maz-ui/compare/v4.0.0-beta.10...v4.0.0-beta.11)